backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
qdrant-client
sqlmodel
psycopg2-binary
